requirements.txt
streamlit
numpy
faiss-cpu
langchain
openai
python-dotenv
